GetState does fetch and remote state last
The fetch is the only operation that is likely to fail for repos that
don't yet have remotes or if the remote cannot be reached.

If that happens, all other info on the repo will be acquired.

There is no need to panic(), just print an error message when querying a
single repo.