Rewrite recursion code
It was completely silly before:

Calling the function once with recurse=True but internally it calls
itself recursively with args.recurse.  That made it recurse at least
once but it was super confusing.

This way it checks the value of 'recurse' more than needed but at least
it's not confusing.